+++ src/Eagle.hpp
@@ -172,9 +172,9 @@ namespace EAGLE {
     double findLongHapMatches(uint64 n0, uint64 nF1, uint64 nF2, int iter);
     double runHMM(uint64 n0, uint64 nF1, uint64 nF2, int iter, uint beamWidth, uint maxHapStates);
     std::vector <bool> computeRefIsMono(const std::vector <uint> &bestHaps) const;
-    float runPBWT(uint64 n0, uint64 nF1, uint64 nF2, int Kpbwt, float cMexpect, double histFactor,
+    float runPBWT(uint64 n0, uint64 nF1, uint64 nF2, int Kpbwt, double cMexpect, double histFactor,
 		  bool runReverse, bool useTargetHaps, bool impMissing);
-    float runPBWT(uint64 n0, uint64 nF1, uint64 nF2, int Kpbwt, float cMexpect, double histFactor,
+    float runPBWT(uint64 n0, uint64 nF1, uint64 nF2, int Kpbwt, double cMexpect, double histFactor,
 		  bool runReverse, bool useTargetHaps, bool impMissing, int usePS,
 		  const std::vector < std::pair <int, int> > &conPS);
     void imputeMissing(const HapHedge &hapHedge, uint64 n0);
