+++ src/EagleParams.hpp
@@ -51,7 +51,7 @@ namespace EAGLE {
     bool pbwtOnly; // in non-ref mode, don't run Steps 1 or 2
     int runStep2; // in non-ref mode, do/don't run Step 2
     int pbwtIters;
-    float expectIBDcM; // expected length of an IBD segment (for transition probabilities)
+    double expectIBDcM; // expected length of an IBD segment (for transition probabilities)
     double histFactor; // history length multiplier
 
     // QC params
